Progress commit: refactoring partnership DTOs, resolves, services
