--- src/components/partnership/agreement-status.ts
@@ -1,11 +1,0 @@
-import { registerEnumType } from 'type-graphql';
-
-export enum PartnershipAgreementStatus {
-  NotAttached = 'na',
-  AwaitingSignature = 'as',
-  Signed = 's',
-}
-
-registerEnumType(PartnershipAgreementStatus, {
-  name: 'PartnershipAgreementStatus',
-});
