+++ src/components/partnership/dto/list-partnership.dto.ts
@@ -1,4 +1,4 @@
-import { Field, ID, InputType, ObjectType } from 'type-graphql';
+import { Field, InputType, ObjectType } from 'type-graphql';
 import {
   PaginatedList,
   SecuredList,
