--- src/components/partnership/partnership-type.ts
@@ -1,11 +1,0 @@
-import { registerEnumType } from 'type-graphql';
-
-export enum PartnershipType {
-  Managing = 'm',
-  Funding = 'f',
-  Impact = 'i',
-  Technical = 't',
-  Resource = 'r',
-}
-
-registerEnumType(PartnershipType, { name: 'PartnershipType' });
